Merge pull request #21 from thefundamentaltheor3m/E8
We completely overhauls the E8 file by introducing two definitions of E8:
`E8_Matrix` and `E8_Set`. We then prove they are equivalent, as well as several
feature about them, such as that the vectors of `E8_Matrix` are linearly
independent.

We also bump Mathlib version and modifies the blueprint to reflect the changes.

Co-authored-by: Sidharth Hariharan <sidharth.hariharan21@imperial.ac.uk>
Co-authored-by: Gareth Ma <grhkm21@gmail.com>