chore: add `create-release` workflow (#95)
This workflow file uses the new `lean-release-tag action` which creates
a Git tag when your project updates to a new Lean release.

It works by looking for changes to the `lean-toolchain` file in the Git
history. For every new release that is found in this file, a tag is
created on the first commit.

For further details, you might want to see the action [README
file](https://github.com/leanprover-community/lean-release-tag).

Co-authored-by: Sidharth Hariharan <sidharth.hariharan21@imperial.ac.uk>