chore: use `lakefile.toml` config file (#98)
Automatically translated by running `lake translate-config toml`.

TOML config files are simpler to configure and maintain for the vast
majority of Lean projects. The TOML config format is fully declarative
and since TOML parsers are available for a wide variety of languages,
using this format facilitates integration with tools that are not
written in Lean.