Restrict to imaginary axis (#116)
Closes #87 

This PR is two-fold:

- Define some APIs for restriction to imaginary axis, and state relevant
theorems without proofs.
- Replace previous `Delta_im_line_im_part` with `Delta_imag_axis_real`
and `Delta_imag_axis_pos`, which is the actual theorems that we need.
The proofs are incomplete but won't be hard.

After this PR, I'd put definitions related to the limits of the
restricted functions as $t \to 0^+$, and state auxiliary lemmas on
"magic modular forms".

---------

Co-authored-by: Sidharth Hariharan <sidharthhariharan@cmu.edu>